feat: create user access token

+++ internal/module/user/service/user_setting_test.go
@@ -27,7 +27,7 @@ func TestService_UpsertAccessToken(t *testing.T) {
 		UpsertUserSetting(ctx, gomock.Any()).
 		Return(nil)
 
-	userId := 1
+	userId := int32(1)
 	accessToken := "dsafadsfasdfadsfasdf"
 	description := "登录"
